% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_image.R
\name{bingmap_image_ggplot}
\alias{bingmap_image_ggplot}
\title{Get map image and plot with magick::image_ggplot()}
\usage{
bingmap_image_ggplot(..., interpolate = FALSE)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_map_image]{get_map_image}}
  \describe{
    \item{\code{location}}{A \code{sf}, \code{sfc}, or \code{bbox} object (centroid point is used for
location) or numeric vector in format \code{c(latitude, longitude)}. Defaults to
\code{NULL}.}
    \item{\code{query}}{String with query for location. query is ignored if a location
is provided.  Defaults to \code{NULL}.}
    \item{\code{imagery}}{String with imagery type, Default: 'BirdsEye' Supported values
include:
\itemize{
\item Aerial: Aerial imagery.
\item AerialWithLabels: Aerial imagery with a road overlay.
\item AerialWithLabelsOnDemand: Aerial imagery with on-demand road overlay.
\item Streetside: Street-level imagery.
\item BirdsEye: Birds Eye (oblique-angle) imagery.
\item BirdsEyeWithLabels: Birds Eye (oblique-angle) imagery with a road overlay.
\item Road: Roads without additional imagery.
\item CanvasDark: A dark version of the road maps.
\item CanvasLight: A lighter version of the road maps which also has some of the
details such as hill shading disabled.
\item CanvasGray: A grayscale version of the road maps
}}
    \item{\code{width,height,mapsize}}{Width and height in pixels or use mapsize to
provide a vector of c(width, height). If mapsize is provided, width and
height are ignored. Default: 600px width, 400px height, mapsize is \code{NULL}.}
    \item{\code{zoom}}{Numeric vector between 0 and 20 for imagery other than Bird's Eye
maps or 18 to 22 for Bird's Eye maps. Default: 18}
    \item{\code{orientation}}{Orientation as a character string ("N", "E", "S", "W") or
length 1 numeric vector (0,90,180,270). Other numeric orientations (from
-360 to 720) are matched to the closest value, e.g. 35 to 0 or 75 to 90.
Default: 0}
    \item{\code{nudge}}{Numeric vector in the format, \verb{c(meters to shift latitude, meters to shift longitude)}, e.g. \code{c(100, 0)} to shift center 100 meters in
latitude.  Defaults to \code{NULL}.}
    \item{\code{key}}{Bing Maps API Key, Default: \code{Sys.getenv("BING_MAPS_API_KEY")}}
    \item{\code{check}}{If \code{TRUE}, check the map metadata which returns an error if the
image is unavailable. Default \code{FALSE}.}
    \item{\code{strip}}{drop image comments and metadata}
  }}

\item{interpolate}{passed to \link[ggplot2:annotation_raster]{ggplot2::annotation_raster}}
}
\description{
Wrapper for \code{\link[magick:image_ggplot]{magick::image_ggplot()}} where all ... parameters are passed to
\code{\link[=get_map_image]{get_map_image()}}.
}
