% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{geocode}
\usage{
geocode(
  addressLine = NULL,
  locality = NULL,
  adminDistrict = NULL,
  postalCode = NULL,
  countryRegion = NULL,
  includeNeighborhood = NULL
)
}
\arguments{
\item{addressLine}{Street address of the location (e.g. "123 Market Street")}

\item{locality}{City/neighborhood of the location (e.g. "San Francisco")}

\item{adminDistrict}{Subdivision of the location, such as the state/province (e.g. "CA")}

\item{postalCode}{Postal/ZIP code of the location (e.g. 94114)}

\item{countryRegion}{ISO country code of the location (e.g. US)}

\item{includeNeighborhood}{Whether the response should include the neighborhood of the location (e.g. "0" or "1")}
}
\value{
S3 object containing the API response. Includes: \cr
"chords": Named list containing the geocoded coordinates\cr
"content": Complete parsed response, including the geocoded coordinates as well as other information about the location\cr
"params": Parameters inputted into the API from the user\cr
"response": Response object returned from httr, including the request URL, status code returned, and time of request\cr
}
\description{
Geocodes an address using the Bing Maps Locations API
For more information, please see https://docs.microsoft.com/en-us/bingmaps/rest-services/locations/find-a-location-by-address
}
