% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode_csv.R
\name{geocode_csv}
\alias{geocode_csv}
\title{geocode_csv}
\usage{
geocode_csv(
  csv_path,
  addressLine_col = NULL,
  locality_col = NULL,
  adminDistrict_col = NULL,
  postalCode_col = NULL,
  countryRegion_col = NULL
)
}
\arguments{
\item{csv_path}{File path of the CSV file with addresses to geocode}

\item{addressLine_col}{Name of the column in the CSV containing the addressLine of the locations}

\item{locality_col}{Name of the column in the CSV containing the locality of the locations}

\item{adminDistrict_col}{Name of the column in the CSV containing the adminDistrict of the locations}

\item{postalCode_col}{Name of the column in the CSV containing the postalCode of the locations}

\item{countryRegion_col}{Name of the column in the CSV containing the countryRegion of the locations}
}
\value{
Tibble containing the original CSV with the columns "latitude" and "longitude" appended \cr
Saves a copy of the inputted CSV including the geocoded coordinates to the same location as the original CSV with '_geocoded' appended to the filename
}
\description{
Geocodes a set of addresses from a CSV file using the geocode function from bingmapr
}
