% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_recognition.R
\name{location_recognition}
\alias{location_recognition}
\title{location_recognition}
\usage{
location_recognition(
  lat,
  long,
  radius = 0.25,
  top = 5,
  dateTime = NULL,
  type = NULL,
  distanceUnit = "mi",
  includeEntityTypes = NULL,
  verboseplacenames = NULL
)
}
\arguments{
\item{lat}{Latitude of the location (e.g. '37.7708' )}

\item{long}{Longitude of the location (e.g. '-122.4195')}

\item{radius}{Radius (using units specified in distanceUnit) in which to find locations (e.g. '1.5'), default value is 0.25, maximum is 2}

\item{top}{Number of nearby entities to include (e.g. '10'), default value is 5, maximum is 20}

\item{dateTime}{Only include entities open at the specified date and time (e.g. '2021-01-01 12:00:00')}

\item{type}{Only include entities with the specified category types (e.g. 'EatDrink')}

\item{distanceUnit}{Units used for the radius parameter (e.g. 'km' or 'mi'), default value is 'mi'}

\item{includeEntityTypes}{Only include the specified entity types (e.g. 'naturalPOI')}

\item{verboseplacenames}{Whether the returned location names should be represented with their official abbreviations or in expanded form (e.g. 'true' or 'false'), default value is 'false'}
}
\value{
S3 object containing the API response. Includes: \cr
"business": Business entities returned, including their name, address, website, phone number, etc.\cr
"naturalPOI": Natural POIs returned, including their name and type\cr
"params": Parameters inputted into the API from the user\cr
"response": Response object returned from httr, including the request URL, status code returned, and time of request\cr
}
\description{
Returns a list of entities (business entities and natural points of interest) near an inputted coordinate pair using the Bing Maps Locations API
For more information, please see https://docs.microsoft.com/en-us/bingmaps/rest-services/locations/location-recognition
}
