% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_geocode.R
\name{reverse_geocode}
\alias{reverse_geocode}
\title{reverse_geocode}
\usage{
reverse_geocode(
  lat,
  long,
  includeEntityTypes = NULL,
  verboseplacenames = NULL,
  includeNeighborhood = NULL
)
}
\arguments{
\item{lat}{Latitude of the location (e.g. '37.7708' )}

\item{long}{Longitude of the location (e.g. '-122.4195')}

\item{includeEntityTypes}{Only include the specified entity types (e.g. 'naturalPOI')}

\item{verboseplacenames}{Whether the returned location names should be represented with their official abbreviations or in expanded form (e.g. 'true' or 'false'), default value is 'false'}

\item{includeNeighborhood}{Whether the response should include the neighborhood of the location (e.g. "0" or "1")}
}
\value{
S3 object containing the API response. Includes: \cr
"address": Complete geocoded address\cr
"content": Complete parsed response, including the geocoded address separated by descriptor as well as other information about the location\cr
"params": Parameters inputted into the API from the user\cr
"response": Response object returned from httr, including the request URL, status code returned, and time of request\cr
}
\description{
Reverse geocodes an address using the Bing Maps Locations API
For more information, please see https://docs.microsoft.com/en-us/bingmaps/rest-services/locations/find-a-location-by-point
}
