% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_image.R
\name{bing_static_map}
\alias{bing_static_map}
\alias{get_request_url}
\alias{get_map_image}
\alias{get_map_meta}
\title{Get and plot Bing Static Maps}
\usage{
get_request_url(
  location = NULL,
  query = NULL,
  imagery = "BirdsEye",
  width = 600,
  height = 400,
  mapsize = NULL,
  zoom = 18,
  orientation = 0,
  nudge = NULL,
  key = Sys.getenv("BING_MAPS_API_KEY"),
  check = FALSE
)

get_map_image(
  location = NULL,
  query = NULL,
  imagery = "BirdsEye",
  width = 600,
  height = 400,
  mapsize = NULL,
  zoom = 18,
  orientation = 0,
  nudge = NULL,
  key = Sys.getenv("BING_MAPS_API_KEY"),
  check = TRUE,
  strip = TRUE
)

get_map_meta(
  location = NULL,
  query = NULL,
  imagery = "BirdsEye",
  width = 600,
  height = 400,
  mapsize = NULL,
  zoom = 18,
  orientation = 0,
  nudge = NULL,
  key = Sys.getenv("BING_MAPS_API_KEY"),
  bbox = FALSE
)
}
\arguments{
\item{location}{A sf object (centroid point is used for location) or numeric
vector in format c(latitude, longitude)}

\item{query}{String with query for location. Ignored if a location is
provided.}

\item{imagery}{String with imagery type, Default: 'BirdsEye' Supported values
include:
\itemize{
\item Aerial: Aerial imagery.
\item AerialWithLabels: Aerial imagery with a road overlay.
\item AerialWithLabelsOnDemand: Aerial imagery with on-demand road overlay.
\item Streetside: Street-level imagery.
\item BirdsEye: Birds Eye (oblique-angle) imagery.
\item BirdsEyeWithLabels: Birds Eye (oblique-angle) imagery with a road overlay.
\item Road: Roads without additional imagery.
\item CanvasDark: A dark version of the road maps.
\item CanvasLight: A lighter version of the road maps which also has some of the
details such as hill shading disabled.
\item CanvasGray: A grayscale version of the road maps
}}

\item{width, height, mapsize}{Width and height in pixels or use mapsize to
provide a vector of c(width, height). If mapsize is provided, width and
height are ignored. Default: 600px wide and 400px high}

\item{zoom}{Numeric vector between 0 and 20 for imagery other than Bird's Eye
maps or 18 to 22 for Bird's Eye maps. Default: 18}

\item{orientation}{Orientation as a character string ("N", "E", "S", "W") or
length 1 numeric vector (0,90,180,270). Other numeric orientations (from
-360 to 720) are matched to the closest value, e.g. 35 to 0 or 75 to 90.
Default: 0}

\item{nudge}{Numeric vector in the format, c(meters to shift latitude, meters
to shift longitude), e.g. c(100, 0) to shift center 100 meters in latitude}

\item{key}{Bing Maps API Key, Default: Sys.getenv("BING_MAPS_API_KEY")}

\item{check}{If \code{TRUE}, check the map metadata which returns an error if the
image is unavailable (Default \code{FALSE} for get_request_url and \code{TRUE} for
get_map_image)}

\item{strip}{drop image comments and metadata}

\item{bbox}{If \code{TRUE}, return a bbox class object based on the bounding box
values from the map metadata. If \code{FALSE} (default), return the full JSON
metadata. (get_map_meta only)}
}
\value{
get_request_url returns the request URL for the Static Map API

get_map_image returns an image from \code{magick::image_read}

get_map_meta returns the JSON with the map metadata or a bbox for the
map area
}
\description{
See the documentation on Bing Static Maps for reference:
https://docs.microsoft.com/en-us/bingmaps/rest-services/imagery/get-a-static-map
}
\details{
Get API key from https://www.bingmapsportal.com/
}
