% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_image.R
\name{bing_static_map}
\alias{bing_static_map}
\alias{get_request_url}
\alias{get_map_image}
\alias{plot_map_image}
\title{Get and plot Bing Static Maps}
\usage{
get_request_url(
  location = NULL,
  query = NULL,
  imagery = "BirdsEye",
  mapsize = c(600, 400),
  zoom = 18,
  orientation = 0,
  key = Sys.getenv("BING_MAPS_TOKEN")
)

get_map_image(
  location = NULL,
  query = NULL,
  imagery = "BirdsEye",
  mapsize = c(600, 400),
  zoom = 18,
  orientation = 0,
  key = Sys.getenv("BING_MAPS_TOKEN")
)

plot_map_image(
  location = NULL,
  query = NULL,
  imagery = "BirdsEye",
  mapsize = c(600, 400),
  zoom = 18,
  orientation = 0,
  key = Sys.getenv("BING_MAPS_TOKEN")
)
}
\arguments{
\item{location}{PARAM_DESCRIPTION, sf object (centroid point is used for location) or numeric vector in format c(latitude, longitude)}

\item{query}{String with query for location. Ignored if a location is provided.}

\item{imagery}{String with imagery type, Default: 'BirdsEye' Supported values include:
\itemize{
\item Aerial: Aerial imagery.
\item AerialWithLabels: Aerial imagery with a road overlay.
\item AerialWithLabelsOnDemand: Aerial imagery with on-demand road overlay.
\item Streetside: Street-level imagery.
\item BirdsEye: Birds Eye (oblique-angle) imagery.
\item BirdsEyeWithLabels: Birds Eye (oblique-angle) imagery with a road overlay.
\item Road: Roads without additional imagery.
\item CanvasDark: A dark version of the road maps.
\item CanvasLight: A lighter version of the road maps which also has some of the details such as hill shading disabled.
\item CanvasGray: A grayscale version of the road maps
}}

\item{mapsize}{Numeric vector with width and height in pixels, Default: c(600, 400) 600px wide, 400px high}

\item{zoom}{Numeric vector between 0 and 20 for imagery other than Bird's Eye maps or 18 to 22 for Bird's Eye maps. Default: 18}

\item{orientation}{Character string with orientation  ("N", "E", "S", "W") or numeric direction (0,90,180,270), Default: 0}

\item{key}{Bing Maps API Key, Default: Sys.getenv("BING_MAPS_TOKEN")}
}
\value{
get_request_url returns the request URL for the Static Map API

get_map_image returns a raster array with JPEG file/content for the map

plot_map plots the static map image
}
\description{
See the documentation on Bing Static Maps for reference: https://docs.microsoft.com/en-us/bingmaps/rest-services/imagery/get-a-static-map
}
\details{
Get API key from https://www.bingmapsportal.com/
}
